setup Typescript project structure ready for chrome browser extension development
This is the first substantial commit towards building the browser extension.
Here we create a Typescript project that can do the minimal actions
required of a chrome extension; console.log'ing to the browser Console
on page load, and console.log to the extension's service worker.

Next we start adding actual headline fetching to the contentScript.ts