Add logic in contentScript for locating and updating headline text with a simple color and case change. Also add in a basic Deno + Hono backend in back/ which we'll improve later
This commit sets up a minimal working implementation of a browser extension that detects a headline
HTML element and then changes it (right now it simply changes the text color to blue, and uppercases
everything). This is still very brittle, it doesn't work on some pages @melvillian tried (for instance,
this page: https://www.cbc.ca/news/canada/ottawa/lebanese-canadians-fear-for-family-back-home-as-violence-escalates-1.7337299)
but it does work for this page: https://www.msnbc.com/rachel-maddow-show/maddowblog/matters-trump-now-wants-prosecute-harris-rcna173278

@melvillian initially tried to get the backend running locally and returning data as well, but he ran into some problems and
the PR was already so big that he commented out the section that communicates with the background service worker, which
in turn sends AJAX requests to the backend server. So for now it's just doing a dummy headline transform.

In future PRs we will need to:
* add correct interaction with the backend
* make the backend take both headline and article content, and communicate with an LLM to decide whether to change the headline or not
* fix all of the TODOs